pyproject.toml
[project]
name = "knowledge-capture"
version = "0.1.0"
description = "High-End RAG API with Voyage and Gemini"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "beanie>=1.25.0",
    "motor>=3.3.0",
    "pydantic-settings>=2.1.0",
    "docling>=2.4.0",
    "voyageai>=0.2.0",
    "google-generativeai>=0.4.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0"
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.3.0"
]
